@Wyverald 2024-01-17
this is superseded by https://github.com/bazelbuild/bazel/commit/d9169ab4deca6d533efe7b5f99c281d58cbeb638